grid-redirects: moved to earlier call
