+++ pkg/interface/src/logic/lib/history.ts
@@ -1,0 +1,7 @@
+import { createBrowserHistory } from 'history';
+
+const history = createBrowserHistory({
+  basename: '/apps/landscape'
+});
+
+export default history;
