+++ pkg/interface/src/logic/reducers/metadata-update.ts
@@ -31,6 +31,7 @@ const associations = (json: MetadataUpdate, state: MetadataState): MetadataState
 
     state.associations = metadata;
     state.loaded = true;
+    state.onLoad();
   }
   return state;
 };
