+++ pkg/interface/src/logic/state/metadata.ts
@@ -1,12 +1,14 @@
 import { Association, Associations, MetadataUpdatePreview } from '@urbit/api';
 import _ from 'lodash';
 import { useCallback, useEffect, useState } from 'react';
+import ob from 'urbit-ob';
 import {
   createState,
   createSubscription,
   reduceStateN
 } from './base';
 import airlock from '~/logic/api';
+import history from '~/logic/lib/history';
 import { reduce } from '../reducers/metadata-update';
 
 export const METADATA_MAX_PREVIEW_WAIT = 150000;
@@ -16,6 +18,7 @@ export interface MetadataState {
   loaded: boolean;
   getPreview: (group: string) => Promise<MetadataUpdatePreview
   >;
+  onLoad: () => void;
   previews: {
     [group: string]: MetadataUpdatePreview
   }
@@ -53,6 +56,9 @@ const useMetadataState = createState<MetadataState>(
         }
         throw e;
       }
+    },
+    onLoad: () => {
+      handleGridRedirect();
     }
   }),
   ['loaded'],
@@ -66,6 +72,12 @@ const useMetadataState = createState<MetadataState>(
       })
   ]
 );
+
+const { graph, groups } = useMetadataState.getState().associations;
+
+if (Object.keys(graph).length > 0 || Object.keys(groups).length > 0) {
+  handleGridRedirect();
+}
 
 export function useAssocForGraph(graph: string) {
   return useMetadataState(
@@ -109,6 +121,99 @@ export function usePreview(group: string) {
   const preview = previews[group];
 
   return { error, preview };
+}
+
+function getGroupResourceRedirect(key: string) {
+  const association = useMetadataState.getState().associations.graph[`/ship/${key}`];
+  const { metadata } = association;
+  if(!association || !('graph' in metadata.config)) {
+    return '';
+  }
+  return `/~landscape${association.group}/resource/${metadata.config.graph}${association.resource}`;
+}
+
+function getPostRedirect(key: string, segs: string[]) {
+  const association = useMetadataState.getState().associations.graph[`/ship/${key}`];
+  const { metadata } = association;
+  if(!association || !('graph' in metadata.config)) {
+    return '';
+  }
+  return `/~landscape${association.group}/feed/thread/${segs.slice(0, -1).join('/')}`;
+}
+
+function getChatRedirect(chat: string, segs: string[]) {
+  const qs = segs.length > 0 ? `?msg=${segs[0]}` : '';
+  return `${getGroupResourceRedirect(chat)}${qs}`;
+}
+
+function getPublishRedirect(graphKey: string, segs: string[]) {
+  const base = getGroupResourceRedirect(graphKey);
+  if(segs.length === 3) {
+    return `${base}/note/${segs[0]}`;
+  } else if (segs.length === 4) {
+    return `${base}/note/${segs[0]}?selected=${segs[2]}`;
+  }
+  return base;
+}
+
+function getLinkRedirect(graphKey: string, segs: string[]) {
+  const base = getGroupResourceRedirect(graphKey);
+  if(segs.length === 1) {
+    return `${base}/index/${segs[0]}`;
+  } else if (segs.length === 3) {
+    return `${base}/index/${segs[0]}?selected=${segs[1]}`;
+  }
+  return base;
+}
+
+function getGraphRedirect(link: string) {
+  const [,mark, ship, name, ...rest] = link.split('/');
+  const graphKey = `${ship}/${name}`;
+  switch(mark) {
+    case 'graph-validator-dm':
+      return `/~landscape/messages/dm/${ob.patp(rest[0])}`;
+    case 'graph-validator-chat':
+      return getChatRedirect(graphKey, rest);
+    case 'graph-validator-publish':
+      return getPublishRedirect(graphKey, rest);
+    case 'graph-validator-link':
+    return getLinkRedirect(graphKey, rest);
+    case 'graph-validator-post':
+      return getPostRedirect(graphKey, rest);
+    default:
+      return'';
+  }
+}
+
+function getInviteRedirect(link: string) {
+  const [,,app,uid] = link.split('/');
+  return `/invites/${app}/${uid}`;
+}
+
+function getDmRedirect(link: string) {
+  const [,,ship] = link.split('/');
+  return `/~landscape/messages/dm/${ship}`;
+}
+
+function getNotificationRedirect(link: string) {
+  if(link.startsWith('/graph-validator')) {
+    return getGraphRedirect(link);
+  } else if (link.startsWith('/invite')) {
+    return getInviteRedirect(link);
+  } else if (link.startsWith('/dm')) {
+    return getDmRedirect(link);
+  }
+}
+
+function handleGridRedirect() {
+  const query = new URLSearchParams(window.location.search);
+
+  if(query.has('grid-note')) {
+    history.push(getNotificationRedirect(query.get('grid-note')));
+  } else if(query.has('grid-link')) {
+    const link = decodeURIComponent(query.get('grid-link')!);
+    history.push(`/perma${link}`);
+  }
 }
 
 export function useGraphsForGroup(group: string) {
